marketing_system_ingest_pipeline/requirements.txt
Flask==2.1.0
pandas
numpy
jsons
gunicorn==20.1.0
google-cloud-pubsub
google-cloud
google-cloud-storage
google-cloud-bigquery==3.2.0
PyYAML==5.4.1
google-cloud-logging==2.6.0
discord-webhook
